requirements.txt
google-generativeai>=0.3.0
requests>=2.31.0
Pillow>=10.0.0
python-dotenv>=1.0.0
playwright>=1.40.0
pydantic>=2.0.0
InquirerPy>=0.4.0
Flask>=2.2.0
openai>=1.40.0
google-genai>=0.3.0
